split out components into folders to make it easier to read
